+++ los_angeles/src/main/java/org/eqasim/los_angeles/mode_choice/LosAngelesModeAvailability.java
@@ -1,4 +1,4 @@
-package org.eqasim.sao_paulo.mode_choice;
+package org.eqasim.los_angeles.mode_choice;
 
 import java.util.Collection;
 import java.util.HashSet;
@@ -6,11 +6,12 @@ import java.util.List;
 
 import org.matsim.api.core.v01.TransportMode;
 import org.matsim.api.core.v01.population.Person;
+import org.matsim.core.population.PersonUtils;
 
 import ch.ethz.matsim.discrete_mode_choice.model.DiscreteModeChoiceTrip;
 import ch.ethz.matsim.discrete_mode_choice.model.mode_availability.ModeAvailability;
 
-public class SaoPauloModeAvailability implements ModeAvailability {
+public class LosAngelesModeAvailability implements ModeAvailability {
 	@Override
 	public Collection<String> getAvailableModes(Person person, List<DiscreteModeChoiceTrip> trips) {
 		Collection<String> modes = new HashSet<>();
@@ -21,15 +22,19 @@ public class SaoPauloModeAvailability implements ModeAvailability {
 		//modes.add(TransportMode.bike);
 
 		// Check car availability
-		boolean carAvailability = true;
+				boolean carAvailability = true;
 
-		if ("never".equals((String) person.getAttributes().getAttribute("carAvailability"))) {
-			carAvailability = false;
-		}
+				if (PersonUtils.getLicense(person).equals("no")) {
+					carAvailability = false;
+				}
 
-		if (carAvailability) {
-			modes.add(TransportMode.car);
-		}
+				if ("none".equals((String) person.getAttributes().getAttribute("carAvailability"))) {
+					carAvailability = false;
+				}
+
+				if (carAvailability) {
+					modes.add(TransportMode.car);
+				}
 
 		// Add special mode "outside" if applicable
 		Boolean isOutside = (Boolean) person.getAttributes().getAttribute("outside");