+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # CHANGELOG
-
+- Remove bike from Sao Paulo
+- Add Los Angeles and San Francisco
 - Add possibility to choose modes to RunPopulationRouting
 - Bugfix: EnrichedTransitRoutingModule wrote wrong link and facilities in "pt interaction" activities, which lead to problems when cutting the population